+++ portals-ui/sites/ngs-portal/src/pages/project/components/index.ts
@@ -1,0 +1,1 @@
+export { ProjectHeader } from './project-header';
