+++ portals-ui/sites/ngs-portal/src/pages/project/project.tsx
@@ -1,36 +1,70 @@
-import { useEffect } from 'react';
+import { useEffect, useMemo, useState } from 'react';
 import { useParams } from 'react-router';
 import { useProjectsStore } from '../../state/projects/hooks';
 import { loadProjects } from '../../state/projects/load-projects';
-import { Breadcrumb, Spin } from 'antd';
+import { noop } from '@cloud-pipeline/core';
+import { ProjectHeader } from './components';
+import { HomeIcon } from '@heroicons/react/24/outline';
+import { Breadcrumb } from 'antd';
 import { Link } from 'react-router-dom';
-import { HomeIcon } from '@heroicons/react/24/solid';
-import { ItemLayout } from '../../shared/ui';
-import { AppRoutes, RoutePath } from '../../shared/constants/routes';
+import { RoutePath, AppRoutes } from '../../shared/constants/routes';
+import { ItemLayout, PageSpinner } from '../../shared/ui';
 
-export function ProjectPage() {
+export const ProjectPage = () => {
   const { projectId } = useParams();
-  const { projects, error, pending, getProjectById } = useProjectsStore();
+  const {
+    projects,
+    error,
+    pending: isProjectsPending,
+    getProjectById,
+  } = useProjectsStore();
   const project = getProjectById(Number(projectId));
 
+  const [activeTabKey, setActiveTabKey] = useState('info');
+
   useEffect(() => {
-    if (!projects?.length) {
-      loadProjects()
-        .then(() => {})
-        .catch(() => {});
+    if (!projects && !isProjectsPending) {
+      loadProjects().then(noop).catch(noop);
     }
-  }, [projects?.length]);
+  }, [projects, isProjectsPending]);
+
+  const tabs = useMemo(
+    () => [
+      {
+        key: 'info',
+        label: <span className="px-4">Info</span>,
+        content: <div>Info</div>,
+      },
+      {
+        key: 'storage',
+        label: <span className="px-4">Storage</span>,
+        content: <div>Storage</div>,
+      },
+      {
+        key: 'pipelines',
+        label: <span className="px-4">Pipelines</span>,
+        content: <div>Pipelines</div>,
+      },
+      {
+        key: 'history',
+        label: <span className="px-4">History</span>,
+        content: <div>History</div>,
+      },
+    ],
+    [],
+  );
+
+  const activeTab = useMemo(
+    () => tabs.find((tab) => tab.key === activeTabKey)!,
+    [activeTabKey, tabs],
+  );
 
   if (error) {
     return <div>{error}</div>;
   }
 
-  if (pending && !projects?.length) {
-    return (
-      <div className="size-full flex items-center justify-center">
-        <Spin size="large" />
-      </div>
-    );
+  if (isProjectsPending && !projects?.length) {
+    return <PageSpinner />;
   }
 
   if (!project) {
@@ -49,17 +83,22 @@ export function ProjectPage() {
             ),
           },
           { title: <Link to={RoutePath[AppRoutes.PROJECTS]}>Projects</Link> },
-          { title: project.name },
+          { title: project?.name },
         ]}
       />
 
       <ItemLayout
         className="flex-grow"
-        header={<div>{project.name}</div>}
-        main={<div>Main</div>}
-        asideTop={<div>History</div>}
-        asideBottom={<div>Permissions</div>}
+        header={
+          <ProjectHeader
+            project={project}
+            tabs={tabs}
+            onChangeTab={setActiveTabKey}
+            activeKey={activeTabKey}
+          />
+        }
+        main={<div>{activeTab.content}</div>}
       />
     </div>
   );
-}
+};
