+++ portals-ui/sites/ngs-portal/src/shared/ui/index.ts
@@ -1,2 +1,3 @@
 export { SelectFilter } from './select-filter';
 export { ItemLayout } from './item-layout';
+export { PageSpinner } from './page-spinner';
