+++ portals-ui/sites/ngs-portal/src/shared/ui/page-spinner.tsx
@@ -1,0 +1,9 @@
+import { Spin } from 'antd';
+
+export const PageSpinner = () => {
+  return (
+    <div className="size-full flex items-center justify-center">
+      <Spin size="large" />
+    </div>
+  );
+};
