+++ portals-ui/sites/ngs-portal/src/state/projects/hooks.ts
@@ -2,7 +2,6 @@ import type { ProjectsState, ProjectsStore } from './types.ts';
 import { useStore } from 'zustand';
 import { projectsStore } from './store.ts';
 import { useMemo } from 'react';
-import type { Project } from '@cloud-pipeline/core';
 
 export function useProjectsStore(): ProjectsStore {
   return useStore(projectsStore);
@@ -18,18 +17,4 @@ export function useProjectsState(): ProjectsState {
     }),
     [projects, pending, error],
   );
-}
-
-export function useProject(
-  projectId: string | number | undefined,
-): Project | undefined {
-  const { projects } = useProjectsState();
-  return useMemo(() => {
-    if (projectId !== undefined && projects) {
-      return projects.find(
-        (project) => String(project.id) === String(projectId),
-      );
-    }
-    return undefined;
-  }, [projectId, projects]);
 }
